project/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-account==0.1.14
django-allauth==0.58.2
django-cors-headers==4.3.1
django-filter==23.4
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
